refactor automl/common/parameters.py
