+++ pyzoo/zoo/chronos/pipeline/parameters.py
@@ -17,8 +17,3 @@
 # where pipeline saves file by default
 DEFAULT_PPL_DIR = "./saved_pipeline"
 DEFAULT_CONFIG_DIR = "./saved_configs"
-
-# logger default name
-DEFAULT_LOGGER_NAME = "AutoML"
-
-DEFAULT_METRIC_NAME = "reward_metric"
